Support early returns when reducing messages (#33)
It is somewhat common to run into what appear as "[junk
bytes](https://github.com/greglook/clj-pgp/pull/33/commits/1fb4a88e6cfcb31486118dba39ad6020d2f0c90d)"
at the end of a PGP encrypted file. It's possible to mitigate some of
the ensuing errors by binding the `error/*handler*` to ignore a variety
of errors encountered after the `:nth` value is positive (that is, after
we've already found the data packet).

However, this is a somewhat brittle approach and doesn't account for all
cases where "junk bytes" might be found (in particular during
decompression).

A more elegant solution (suggested by @greglook) would be for the
library's caller to return a `reduced` value as soon as the data packet
is found, forcing an early return in the overall, nested reduction.

Today, that isn't supported because an early return in one reduction
(e.g. a `PGPLiteralData` `reduce-message` method) will only force that
one reduction to return and then we'll continue reducing the rest of the
packets in the outer reductions (e.g. a surrounding `PGPCompressedData`
reduction).

This PR adds proper support for early returns by making two changes:
-
[Wrapping](https://github.com/greglook/clj-pgp/pull/33/commits/1d212332ac3f5b819e5d3cf9787a57c0d73dd93d#diff-3b80a935df92bf2fc56a7152f11ec2afcee5223ed30594988e8711e3e7d846b7L440-R440)
intermediate reductions in `preserving-reduced`
This makes sure that if the current accumulator is already `reduced?`,
the "parent" reduction will also see a reduced value (this is done by
nesting an extra `reduced` layer, of which `clojure.core/reduce` will
only remove a single layer)
-
[Frontloading](https://github.com/greglook/clj-pgp/pull/33/commits/d0a465da7b7fef4377aba5ac861fa33daf3f4187)
the `reduced?` base condition of the `loop` inside
`clj-pgp.core/read-objects`
Without this, we'll still try to read the next object off of the stream
_before_ checking to see if we should terminate early. That still leaves
the door open to throw an exception instead of gracefully returning.